♻️ Save edit/view data template modal to its own file
